requirements.txt
assemblyai==0.37.0
datasets==3.4.0
imageio==2.25.1
moviepy==2.0.0.dev2
nltk==3.9.1
numpy==2.2.4
opencv_python==4.11.0.86
Pillow==11.1.0
protobuf==6.30.2
pydub==0.25.1
pysrt==1.1.2
soundfile==0.13.1
SpeechRecognition==3.14.1
streamlit==1.43.2
torch==2.6.0
transformers==4.49.0
